+++ src/ElectionGuard/ElectionGuardAPI/API.cs
@@ -36,7 +36,7 @@ namespace ElectionGuard.SDK.ElectionGuardAPI
         /// <param name="trackerPtr"></param>
         /// <returns></returns>
         [DllImport("electionguard", EntryPoint = "API_EncryptBallot")]
-        internal static extern bool EncryptBallot(ushort[] selections,
+        internal static extern bool EncryptBallot(byte[] selections,
                                                   APIConfig config,
                                                   ref ulong currentNumberOfBallots,
                                                   out ulong ballotIdPtr,
